chg: refactored as requested
