+++ src/controller/file/read.py
@@ -22,7 +22,8 @@ from model.macro import Macro, trigger_factory
 from model.media_assets.asset_loading_factory import load_asset
 from model.media_assets.factory_hint import AssetFactoryObjectHint
 from model.media_assets.registry import clear as clear_media_registry
-from model.ofl.fixture import FixtureDefNotFoundError, load_fixture, make_used_fixture
+from model.ofl.fixture import load_fixture, make_used_fixture
+from model.ofl.fixture_not_found_exception import FixtureDefNotFoundError
 from model.scene import FilterPage
 from model.virtual_filters.vfilter_factory import construct_virtual_filter_instance
 from utility import resource_path
